+++ apps/client/src/app/components/admin-market-data-detail/admin-market-data-detail.module.ts
@@ -1,0 +1,13 @@
+import { CommonModule } from '@angular/common';
+import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
+
+import { AdminMarketDataDetailComponent } from './admin-market-data-detail.component';
+
+@NgModule({
+  declarations: [AdminMarketDataDetailComponent],
+  exports: [AdminMarketDataDetailComponent],
+  imports: [CommonModule],
+  providers: [],
+  schemas: [CUSTOM_ELEMENTS_SCHEMA]
+})
+export class GfAdminMarketDataDetailModule {}