+++ apps/client/src/app/services/data.service.ts
@@ -16,6 +16,8 @@ import {
   Access,
   Accounts,
   AdminData,
+  AdminMarketData,
+  AdminMarketDataDetails,
   Export,
   InfoItem,
   PortfolioChart,
@@ -62,6 +64,23 @@ export class DataService {
 
   public fetchAdminData() {
     return this.http.get<AdminData>('/api/admin');
+  }
+
+  public fetchAdminMarketData() {
+    return this.http.get<AdminMarketData>('/api/admin/market-data');
+  }
+
+  public fetchAdminMarketDataBySymbol(
+    aSymbol: string
+  ): Observable<AdminMarketDataDetails> {
+    return this.http.get<any>(`/api/admin/market-data/${aSymbol}`).pipe(
+      map((data) => {
+        for (const item of data.marketData) {
+          item.date = parseISO(item.date);
+        }
+        return data;
+      })
+    );
   }
 
   public deleteAccess(aId: string) {