+++ libs/common/src/lib/interfaces/index.ts
@@ -1,6 +1,8 @@
 import { Access } from './access.interface';
 import { Accounts } from './accounts.interface';
 import { AdminData } from './admin-data.interface';
+import { AdminMarketDataDetails } from './admin-market-data-details.interface';
+import { AdminMarketData } from './admin-market-data.interface';
 import { Export } from './export.interface';
 import { InfoItem } from './info-item.interface';
 import { PortfolioChart } from './portfolio-chart.interface';
@@ -23,6 +25,8 @@ export {
   Access,
   Accounts,
   AdminData,
+  AdminMarketData,
+  AdminMarketDataDetails,
   Export,
   InfoItem,
   PortfolioChart,