+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added tabs with routing to the admin control panel
+- Added a new tab to manage historical data to the admin control panel
 
 ### Changed